# sort: Make use of ExtendedBigDecimal in -g sorting, then attempt to recover some performance